habia dos componentes con el mismo nombre
